+++ packages/api/models/rules.js
@@ -2,8 +2,8 @@
 
 const cloneDeep = require('lodash/cloneDeep');
 const get = require('lodash/get');
-const merge = require('lodash/merge');
 const set = require('lodash/set');
+const merge = require('lodash/merge');
 
 const awsServices = require('@cumulus/aws-client/services');
 const CloudwatchEvents = require('@cumulus/aws-client/CloudwatchEvents');
@@ -160,8 +160,7 @@ class Rule extends Manager {
 
     updatedRuleItem = await this.updateRuleTrigger(updatedRuleItem, stateChanged, valueUpdated);
 
-    return super.update({ name: original.name }, updatedRuleItem,
-      fieldsToDelete);
+    return super.update({ name: original.name }, updatedRuleItem, fieldsToDelete);
   }
 
   async updateRuleTrigger(ruleItem, stateChanged, valueUpdated) {
@@ -242,7 +241,7 @@ class Rule extends Manager {
     await invoke(process.env.invoke, payload);
   }
 
-  async create(item, createdAt) {
+  async create(item) {
     // make sure the name only has word characters
     const re = /\W/;
     if (re.test(item.name)) {
@@ -257,8 +256,8 @@ class Rule extends Manager {
       newRuleItem.state = 'ENABLED';
     }
 
-    newRuleItem.createdAt = createdAt || Date.now();
-    newRuleItem.updatedAt = Date.now();
+    newRuleItem.createdAt = item.createdAt || Date.now();
+    newRuleItem.updatedAt = item.updatedAt || Date.now();
 
     // Validate rule before kicking off workflows or adding event source mappings
     await this.constructor.recordIsValid(newRuleItem, this.schema, this.removeAdditional);